requirements.txt
python-dateutil;python_version<"3.6"
scrapy>=1.3.1,<2.0;python_version<"3"
scrapy<1.5.0;python_version=="3.3.*"
scrapy>=1.1,<1.8.0;python_version=="3.4.*"
scrapy>=1.1;python_version=="3.5.*"
scrapy>=1.3.1;python_version>"3.5"
scrapy>=1.7.1;python_version>="3.10"
packaging
six
cryptography<2.0;python_version=="3.3.*"
pyOpenSSL<17.3.0;python_version=="3.3.*"
